Change size and capacity to explicit int64

+++ exporter/exporterqueue/queue.go
@@ -28,9 +28,9 @@ type Queue[T any] interface {
 	// It returns ErrQueueIsFull if no space is currently available.
 	Offer(ctx context.Context, item T) error
 	// Size returns the current Size of the queue
-	Size() int
+	Size() int64
 	// Capacity returns the capacity of the queue.
-	Capacity() int
+	Capacity() int64
 	// Read pulls the next available item from the queue along with its index. Once processing is
 	// finished, the index should be called with OnProcessingFinished to clean up the storage.
 	// The function blocks until an item is available or if the queue is stopped.
