+++ exporter/exporterqueue/sized_queue_test.go
@@ -20,25 +20,25 @@ func (s sizerInt) Sizeof(el int) int64 {
 func TestSizedQueue(t *testing.T) {
 	q := newSizedQueue[int](7, sizerInt{})
 	require.NoError(t, q.Offer(context.Background(), 1))
-	assert.Equal(t, 1, q.Size())
-	assert.Equal(t, 7, q.Capacity())
+	assert.Equal(t, int64(1), q.Size())
+	assert.Equal(t, int64(7), q.Capacity())
 
 	require.NoError(t, q.Offer(context.Background(), 3))
-	assert.Equal(t, 4, q.Size())
+	assert.Equal(t, int64(4), q.Size())
 
 	// should not be able to send to the full queue
 	require.Error(t, q.Offer(context.Background(), 4))
-	assert.Equal(t, 4, q.Size())
+	assert.Equal(t, int64(4), q.Size())
 
 	_, el, ok := q.pop()
 	assert.Equal(t, 1, el)
 	assert.True(t, ok)
-	assert.Equal(t, 3, q.Size())
+	assert.Equal(t, int64(3), q.Size())
 
 	_, el, ok = q.pop()
 	assert.Equal(t, 3, el)
 	assert.True(t, ok)
-	assert.Equal(t, 0, q.Size())
+	assert.Equal(t, int64(0), q.Size())
 
 	require.NoError(t, q.Shutdown(context.Background()))
 	_, el, ok = q.pop()
@@ -54,13 +54,13 @@ func TestSizedQueue_DrainAllElements(t *testing.T) {
 	_, el, ok := q.pop()
 	assert.Equal(t, 1, el)
 	assert.True(t, ok)
-	assert.Equal(t, 3, q.Size())
+	assert.Equal(t, int64(3), q.Size())
 
 	require.NoError(t, q.Shutdown(context.Background()))
 	_, el, ok = q.pop()
 	assert.Equal(t, 3, el)
 	assert.True(t, ok)
-	assert.Equal(t, 0, q.Size())
+	assert.Equal(t, int64(0), q.Size())
 
 	_, el, ok = q.pop()
 	assert.False(t, ok)
